test ZK health setp and refactor connection listener health setup
